requirements.txt
flask
pychromecast
Pillow

